backend/requirements.txt
# Django Core (совместимо с Python 3.10+)
Django==5.2.10

# Django REST Framework
djangorestframework==3.16.1

# CORS Headers для работы с React фронтендом
django-cors-headers==4.9.0

# JWT Authentication
djangorestframework-simplejwt==5.5.1

# API docs and filtering
drf-yasg==1.21.10
django-filter==25.1

# Twilio для отправки SMS
twilio==9.10.2

# Database
psycopg2-binary==2.9.10
# mysqlclient>=2.1.0      # Для MySQL (раскомментируйте если нужно)

# Static files in production (Vercel)
whitenoise==6.10.0

# App server for Docker/production
gunicorn==23.0.0

# Установка зависимостей:
# pip install -r requirements.txt
#
# Или с использованием виртуального окружения:
# python -m venv venv
# venv\Scripts\activate  # Windows
# source venv/bin/activate  # Linux/Mac
# pip install -r requirements.txt
